Rebase from master and refactor to clean and clear the thing
