+++ app/src/main/java/fr/utt/if26/mytravel/Activities/MainActivity.java
@@ -3,12 +3,8 @@ package fr.utt.if26.mytravel.Activities;
 import android.content.Intent;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
-import android.util.Log;
 import android.view.View;
 import android.widget.Button;
-
-import java.util.ArrayList;
-import java.util.Iterator;
 
 import fr.utt.if26.mytravel.Config.Bdd;
 import fr.utt.if26.mytravel.DAO.PageDAO;
