+++ app/src/main/java/fr/utt/if26/mytravel/Activities/Page_createActivity.java
@@ -1,4 +1,4 @@
-package fr.utt.if26.mytravel;
+package fr.utt.if26.mytravel.Activities;
 
 import android.content.Intent;
 import android.support.v7.app.AppCompatActivity;
@@ -7,7 +7,10 @@ import android.view.View;
 import android.widget.Button;
 import android.widget.EditText;
 
+import fr.utt.if26.mytravel.Config.Bdd;
 import fr.utt.if26.mytravel.DAO.PageDAO;
+import fr.utt.if26.mytravel.Model.Page;
+import fr.utt.if26.mytravel.R;
 
 public class Page_createActivity extends AppCompatActivity {
     private Bdd database;
