+++ app/src/main/res/layout/activity_page_item.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="fr.utt.if26.mytravel.Page_itemActivity"
+    tools:context="fr.utt.if26.mytravel.Activities.Page_itemActivity"
     android:orientation="vertical">
 
     <EditText
